Create cnn.py (update)
The update to the code involves adding a function to calculate the number of people currently in the system. The function is named count_people_in_system and takes the database connection as an argument.

The function works by executing a SQL query to count the number of rows in the table that contains information about people in the system. The SQL query is as follows:
SELECT COUNT(*) FROM people;

Once the query is executed, the function returns the result as an integer value representing the number of people in the system.

The new function can be called from the main code by simply invoking the function and passing in the database connection object. For example:

num_people = count_people_in_system(db_conn)
print(f"There are currently {num_people} people in the system.")
This will print out a message indicating the current number of people in the system, based on the data stored in the database.